Extract WithImportedSwitch function component
